Issue #12 refactor 패키지 이름 수정
add -> save
